Remove DHT and rename DHT11->DHT
The new DHT11 is more useful than the old DHT implementation, so make it
the official one to use now. This also gets rid of some confusion of
which one should be used.
