+++ lib/grovepi/dht/default_trigger.ex
@@ -1,22 +1,22 @@
-defmodule GrovePi.DHT11.DefaultTrigger do
+defmodule GrovePi.DHT.DefaultTrigger do
   @behaviour GrovePi.Trigger
 
   @moduledoc """
-  This is the default triggering mechanism for DHT11 events. The
+  This is the default triggering mechanism for DHT events. The
   event is `:changed` and includes the trigger state. The trigger state
   for the default trigger is a struct containing `temp` and `humidity`
   properties.
 
   ## Examples
-      iex> GrovePi.DHT11.DefaultTrigger.init([])
-      {:ok, %GrovePi.DHT11.DefaultTrigger.State{temp: 0, humidity: 0}}
-      iex> GrovePi.DHT11.DefaultTrigger.update({0, 0}, %{temp: 0, humidity: 0})
+      iex> GrovePi.DHT.DefaultTrigger.init([])
+      {:ok, %GrovePi.DHT.DefaultTrigger.State{temp: 0, humidity: 0}}
+      iex> GrovePi.DHT.DefaultTrigger.update({0, 0}, %{temp: 0, humidity: 0})
       {:ok, %{temp: 0, humidity: 0}}
-      iex> GrovePi.DHT11.DefaultTrigger.update({11.3, 45.5}, %{temp: 0, humidity: 0})
+      iex> GrovePi.DHT.DefaultTrigger.update({11.3, 45.5}, %{temp: 0, humidity: 0})
       {:changed, %{temp: 11.3, humidity: 45.5}}
-      iex> GrovePi.DHT11.DefaultTrigger.update({11.3, 45.5}, %{temp: 11.3, humidity: 45.5})
+      iex> GrovePi.DHT.DefaultTrigger.update({11.3, 45.5}, %{temp: 11.3, humidity: 45.5})
       {:ok, %{temp: 11.3, humidity: 45.5}}
-      iex> GrovePi.DHT11.DefaultTrigger.update({22.5, 34.5}, %{temp: 11.3, humidity: 45.5})
+      iex> GrovePi.DHT.DefaultTrigger.update({22.5, 34.5}, %{temp: 11.3, humidity: 45.5})
       {:changed, %{temp: 22.5, humidity: 34.5}}
   """
