+++ test/grovepi/dht/default_trigger_test.exs
@@ -1,0 +1,4 @@
+defmodule GrovePi.DHT.DefaultTriggerTest do
+  use ExUnit.Case, async: true
+  doctest GrovePi.DHT.DefaultTrigger
+end
