--- test/grovepi/dht_11/default_trigger_test.exs
@@ -1,4 +1,0 @@
-defmodule GrovePi.DHT11.DefaultTriggerTest do
-  use ExUnit.Case, async: true
-  doctest GrovePi.DHT11.DefaultTrigger
-end
